+++ src/UI_music/Admin_UI.java
@@ -58,22 +58,28 @@ import javafx.stage.StageStyle;
  */
 public class Admin_UI extends UI{
     
-    LocalDate dOB;
-    File user = new File("src/data/user.dat");
-    boolean dateSet = false;   
-    ArrayList<Account> listAccount = new ArrayList<>(); 
-    ArrayList<Account> addAccount = new ArrayList<>();
+    private LocalDate dOB;
+    private File user = new File("src/data/user.dat");
+    private boolean dateSet = false;   
+    private ArrayList<Account> listAccount = new ArrayList<>(); 
+    private ArrayList<Account> addAccount = new ArrayList<>();
     
-    ObservableList<Account> list = null; 
+    private ObservableList<Account> list = null; 
     
-    SearchSystem searchSystemMain = new SearchSystem();
-    SearchSystemAccount searchAccount = new SearchSystemAccount();
+    private SearchSystem searchSystemMain = new SearchSystem();
+   private  SearchSystemAccount searchAccount = new SearchSystemAccount();
 
-    TableView<Account> table;
+    private TableView<Account> table;
     
-
-    public Admin_UI(Stage stage) {
+    private Account userAccount;
+    
+    private ReadWriteFile file = new ReadWriteFile();
+    
+    public Admin_UI(Stage stage, Account userAccount) {
         super(stage);
+        
+        this.userAccount = userAccount;
+        
         Scene scene = new Scene(allPane2(), 1280, 960);
         String stylrSheet = getClass().getResource("/style_css/style.css").toExternalForm();
         String stylrSheet2 = getClass().getResource("/style_css/styleAdmin.css").toExternalForm();
@@ -143,7 +149,7 @@ public class Admin_UI extends UI{
         addAccountBtn.setLayoutX(290);
         addAccountBtn.setLayoutY(675); 
         addAccountBtn.setOnAction(e -> {
-            register(null);
+            register();
             refreshTable();
         });
         
@@ -300,7 +306,12 @@ public class Admin_UI extends UI{
         return hBox;
     }
     
-    public static VBox totalPane;
+    @Override
+    public BorderPane myAccount() { 
+        return new Profile(userAccount).getMainPane();
+    }
+    
+    private static VBox totalPane;
     private ScrollPane  AllSong(){
         
         ScrollPane scrollPane = new ScrollPane();
@@ -328,7 +339,7 @@ public class Admin_UI extends UI{
     static Label selectArtist = new Label(""); 
     static ImageView selectImage;
     
-    public static TilePane updateScrollPane(String text){
+    private static TilePane updateScrollPane(String text){
         VBox paneContent;
         Button contentButton;
         ImageView imageView;
@@ -386,8 +397,8 @@ public class Admin_UI extends UI{
     }
     
     //UPDATE CLICKPANE // RUN ONLY ONCE THE PROGRAM RUN 1 PAGE
-    public static VBox updateVBox;
-    public AnchorPane UpdateClikedPane(){
+    private static VBox updateVBox;
+    private AnchorPane UpdateClikedPane(){
         //Image
         AnchorPane updatePane = new AnchorPane();
         updatePane.setLayoutX(760);
@@ -414,217 +425,10 @@ public class Admin_UI extends UI{
         return updatePane;
     } 
     
-    public void register(String email) {
-        //StringProperty name, surname, mail, password, sex;
-        Stage regisStage = new Stage();
-        regisStage.initModality(Modality.APPLICATION_MODAL);
-
-        Label title = new Label("New Account");
-
-        //Fill name surname username email password
-        TextField nameIn = new TextField();
-        nameIn.setPromptText("Name");
-        TextField surnameIn = new TextField();
-        surnameIn.setPromptText("Surname");
-        TextField usernameIn = new TextField();
-        usernameIn.setPromptText("Username");
-        TextField mailIn = new TextField(email);
-        mailIn.setPromptText("Email e.g. Spookify@gmail.com");
-        TextField passIn = new PasswordField();
-        passIn.setPromptText("New Password");
-        TextField cfPassIn = new PasswordField();
-        cfPassIn.setPromptText("Confirm Password");
-
-        // create a date picker 
-        DatePicker date = new DatePicker();
-        // show week numbers 
-        date.setShowWeekNumbers(false);
-        // when datePicker is pressed 
-        date.setOnAction(e -> {
-            dOB = date.getValue();
-            dateSet = true;
-        });
-
-        //Select Gender
-        Label sexText = new Label("Gender");
-        ToggleGroup sexToggle = new ToggleGroup(); //create radio button group
-        RadioButton male = new RadioButton("Male"); //create radio button
-        RadioButton female = new RadioButton("Female");
-        RadioButton otherRadio = new RadioButton("Not specify");
-        male.setToggleGroup(sexToggle); //Set radio button group
-        female.setToggleGroup(sexToggle);
-        otherRadio.setToggleGroup(sexToggle);
-        sexToggle.selectToggle(male); // Set default = male
-        
-        Label adminText = new Label("Administrive");
-        ToggleGroup adminToggle = new ToggleGroup(); //create radio button group
-        RadioButton userRadio = new RadioButton("User"); //create radio button
-        RadioButton adminRadio = new RadioButton("Admin");
-        userRadio.setToggleGroup(adminToggle); //Set radio button group
-        adminRadio.setToggleGroup(adminToggle);
-        adminToggle.selectToggle(userRadio); // Set default = user
-
-        //FORGOT QUESTION
-        Label qText = new Label("Security Question");
-        ComboBox<String> question = new ComboBox<>();
-        question.getItems().addAll(
-                "What's your first school.",
-                "Your favourite pet's name.",
-                "Your father's name"
-        );
-        question.setPromptText("Select / write a question.");
-        question.setEditable(true); //USER CAN WRITE THEIR OWN QUESTION
-        //FORGOT ANSWER
-        TextField answer = new TextField();
-        answer.setPromptText("Answer");
-
-        Button ok = new Button("OK");
-        ok.setOnAction(e -> {
-            System.out.println("Checking information...");
-
-            String userGender = "";
-
-            //Check sex radioButton which is selected
-            if (male.isSelected()) {
-                userGender = "Male";
-            } else if (female.isSelected()) {
-                userGender = "Female";
-            } else {
-                userGender = "N/A";
-            }
-            
-            boolean isAdmin = false;
-
-            //Check admin radioButton which is selected
-            if (adminRadio.isSelected()) {
-                isAdmin = true;
-            } else if (userRadio.isSelected()) {
-                isAdmin = false;
-            }
-            ArrayList<Account> addAccount = new ArrayList<>();
-
-
-            try {
-                listAccount = readFile(user);
-            } catch (Exception ex) {
-                System.out.println("error: " + ex);
-            }
-            boolean uniqueID = true;
-            //Check already username / email
-            for (Account account : listAccount) {
-                String userId = usernameIn.getText(), emailID = mailIn.getText();
-                String chkUser = account.getUsername(), chkEmail = account.getEmail();
-                if ((userId.equals(chkUser) || emailID.equals(chkEmail))) {
-                    uniqueID = false;
-                    break;
-                }
-            }
-            if (uniqueID == false) {
-                AlertBox.displayAlert("Something went wrong!", "Email / username is already exists.");
-            } //Check comfirm password is the same as password will call error password
-            else if (passIn.getText().equals(cfPassIn.getText())) {
-                //Check if it has Blank Field will call error blank field
-                if (nameIn.getText().isBlank() || passIn.getText().isBlank() || usernameIn.getText().isBlank()
-                        || mailIn.getText().isBlank() || dateSet == false || question.equals(null) || answer.getText().isBlank()) {
-                    AlertBox.displayAlert("Something went wrong!", "Please check all the form.\nAnd make sure it was filled.");
-                } else {
-                    //Check Date of Birth
-                    if (dOB.isAfter(LocalDate.now())) { // checkdate if user born after present
-                        System.out.println("User is come from the future.");
-                        AlertBox.displayAlert("Something went wrong!", "Please check a date field again.\n"
-                                + "Make sure that's your date of birth.");
-                    }
-                    //Check email if it's not will call error email [ see function isEmail for more information]
-                    else if (!isEmail(mailIn.getText())) {
-                        AlertBox.displayAlert("Something went wrong!", "Please use another email.");
-
-                    } else {
-                        try {
-                            addAccount = readFile(user);
-                        } catch (IOException | ClassNotFoundException ex) {
-                            System.out.println("Register readFile " + ex);
-                        }
-
-                        addAccount.add(new Account(nameIn.getText(), surnameIn.getText(), usernameIn.getText(), mailIn.getText(),
-                                passIn.getText(), userGender, dOB, question.getValue(), answer.getText(), isAdmin, new Image("default-profile.png")));
-
-                        try {
-                            writeFile(user, addAccount);
-                            System.out.println("Saving account.");
-                        } catch (IOException ex) {
-                            System.out.println("Register writeFile " + ex);
-                        }
-                        try {
-                            listAccount = readFile(user);
-                        } catch (Exception ex) {
-                            System.out.println("Error: " + ex);
-                        }
-                        AlertBox.displayAlert("Register Complete", "Your account has been saved.\nTry to login now.");
-
-                        System.out.println("Registeraion Complete!\n");
-
-                        regisStage.close();
-                    }
-                }
-            } else {
-                AlertBox.displayAlert("Something went wrong!", "Confirm password is not as same as password.");
-            }
-
-        });
-        
-        Button cancel = new Button("Cancel");
-        cancel.setOnAction(e -> {
-            System.out.println("Canceling Registeration.");
-            regisStage.close();
-        });
-
-        HBox row1 = new HBox(20); //Name Row
-        row1.getChildren().addAll(nameIn, surnameIn);
-        row1.setAlignment(Pos.CENTER);
-
-        HBox row3 = new HBox(20);//Password Row
-        row3.getChildren().addAll(passIn, cfPassIn);
-        row3.setAlignment(Pos.CENTER);
-
-        Label title2 = new Label("Date of birth");
-
-        HBox sexRow = new HBox(20);
-        sexRow.getChildren().addAll(male, female, otherRadio);
-        sexRow.setAlignment(Pos.CENTER);
-        
-        HBox adminRow = new HBox(20);
-        adminRow.getChildren().addAll(userRadio, adminRadio);
-        adminRow.setAlignment(Pos.CENTER);
-
-        HBox row2 = new HBox(20); //Button Row
-        row2.getChildren().addAll(ok, cancel);
-        row2.setAlignment(Pos.CENTER);
-
-        VBox column1 = new VBox(20);
-        column1.setPadding(new Insets(10)); //add gap 10px
-        column1.getChildren().addAll(title, adminRow, row1, usernameIn, mailIn, row3, title2, date, sexText, sexRow, qText, question, answer, row2);
-        column1.setAlignment(Pos.CENTER);
-        Scene regScene = new Scene(column1, 360, 600);
-        regisStage.setScene(regScene);
-        regisStage.setResizable(false);
-        regisStage.setTitle("Registeration.");
-        regisStage.showAndWait();
-
+    private void register() {
+        new Register();
     }
-    
-    //ReadFile Function from Registor
-    private ArrayList<Account> readFile(File file) throws FileNotFoundException, IOException, ClassNotFoundException {
-        ObjectInputStream in = new ObjectInputStream(new FileInputStream(file));
-        return (ArrayList<Account>) in.readObject();
-    }
-    
-    //WriteFile Function from Registor
-    private void writeFile(File file, ArrayList<Account> listAccount) throws FileNotFoundException, IOException {
-        ObjectOutputStream out = new ObjectOutputStream(new FileOutputStream(file));
-        out.writeObject(listAccount);
-        out.close();
-    }
-    
+        
     private void refreshTable(){ //get.list -> sorted
         //TRY -CATCH FOR EXCEPTION ... NOTHING TO DO WITH IT
         try {
@@ -649,7 +453,7 @@ public class Admin_UI extends UI{
         
         ArrayList<Account> oldAccounts = new ArrayList<>();
         ArrayList<Account> presentAccounts = new ArrayList<>();
-        oldAccounts = readFile(user);   
+        oldAccounts = file.readFile(user);   
         
         for (Account account : oldAccounts) {
             String chkUser = account.getUsername();
@@ -662,7 +466,7 @@ public class Admin_UI extends UI{
             }
         }
         
-        writeFile(user, presentAccounts);
+        file.writeFile(user, presentAccounts);
 
     }
     
@@ -672,11 +476,6 @@ public class Admin_UI extends UI{
     private boolean isEmail(String email) {
         return email.contains("@") && email.contains(".") && !(email.contains(" ") || email.contains(";") || email.contains(",") || email.contains("..")
                 || email.length() <= 12 || email.length() > 64);
-    }
-
-    @Override
-    public BorderPane myAccount() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
     
 }
