+++ src/UI_music/ReadWriteFile.java
@@ -1,0 +1,39 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package UI_music;
+
+import Component_Music.Account;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
+import java.util.ArrayList;
+
+/**
+ *
+ * @author HCARACH
+ */
+public class ReadWriteFile {
+
+    public ReadWriteFile() {
+    }
+    
+    public ArrayList<Account> readFile(File file) throws FileNotFoundException, IOException, ClassNotFoundException {
+        ObjectInputStream in = new ObjectInputStream(new FileInputStream(file));
+        ArrayList<Account> list = (ArrayList<Account>) in.readObject();
+        in.close();
+        return list;
+    }
+
+    public void writeFile(File file, ArrayList<Account> listAccount) throws FileNotFoundException, IOException {
+        ObjectOutputStream out = new ObjectOutputStream(new FileOutputStream(file));
+        out.writeObject(listAccount);
+        out.close();
+    }
+}
