+++ src/UI_music/User_UI.java
@@ -290,62 +290,8 @@ public class User_UI extends UI {
     Button cancelbt;
         
     @Override
-    public BorderPane myAccount() {
-        
-        BorderPane accountPane = new BorderPane();
-        BorderPane mainPane = new BorderPane();
-        MyAccount myAccount = new MyAccount(userAccount);
-                
-        VBox head = new VBox(10);
-        head.setPadding(new Insets(0,10,20,20));
-        head.getChildren().add(new Text("MY ACCOUNT"));
-        
-         HBox bottom = new HBox(10);
-        bottom.setPadding(new Insets(20,20,0,0));
-        bottom.setAlignment(Pos.CENTER_RIGHT);
-        
-        savebt = new Button("Save");
-        savebt.setOnAction(event -> {
-            if(myAccount.saveAccount()) {
-                AlertBox.displayAlert("Edit Profile", "Saved.");
-                userAccount = myAccount.getMyAccount();
-                myAccount.showAccount(userAccount);
-                accountPane.setCenter(myAccount.getProfilePane());
-                bottom.getChildren().clear();
-                bottom.getChildren().addAll(editbt);
-            } 
-        });
-        
-        cancelbt = new Button("Cancel");
-        cancelbt.setOnAction(event -> {
-            myAccount.showAccount(userAccount);
-            accountPane.setCenter(myAccount.getProfilePane());
-            bottom.getChildren().clear();
-            bottom.getChildren().addAll(editbt);
-        });
-        
-        editbt = new Button("Edit");
-        VBox right = new VBox(10);
-        right.setPadding(new Insets(20,20,20,20));
-        editbt.setOnAction(event -> {
-            myAccount.editAccount();
-            accountPane.setCenter(myAccount.getProfilePane());
-            bottom.getChildren().clear();
-            bottom.getChildren().addAll(cancelbt, savebt);
-        });
-        
-                bottom.getChildren().add(editbt);
-        
-        accountPane.setTop(head);
-        accountPane.setCenter(myAccount.getProfilePane());
-        accountPane.setPadding(new Insets(50, 50, 50, 50));
-        accountPane.setStyle("-fx-background-color: white");
-        accountPane.setBottom(bottom);
-        
-        mainPane.setCenter(accountPane);
-        mainPane.setPadding(new Insets(50, 100, 0, 100));
-        
-        return mainPane;
+    public BorderPane myAccount() { 
+        return new Profile(userAccount).getMainPane();
     }
 
     public void downloader() {
